requirements.txt
faker==19.3.0
pandas~=2.0.3
requests~=2.31.0
PyYAML==6.0.1
tqdm==4.66.1
setuptools~=68.0.0
numpy==1.25.2
mypy
black

pandas-stubs
types-requests
types-PyYAML
types-tqdm
types-setuptools